Move tests/benches into seperate files
